refactor(zk-explorer): Prefer default exports
