+++ .env.example
@@ -44,6 +44,8 @@ TESTNET_LEADERBOARD_BACKEND_URL=""
 
 # ZK Explorer
 ZK_EXPLORER_GITHUB_CLIENT_ID=""
+ZK_EXPLORER_API_BASE_URL="/api"
 
+# OFAC
 OFAC_COUNTRY_CODES='["<OFAC_COUNTRY_CODE>"]'
 OFAC_REGIONS='["<OFAC_REGION>"]'
