+++ apps/zk-explorer/components/ProjectCard.tsx
@@ -2,7 +2,7 @@ import { StarIcon } from '@radix-ui/react-icons';
 import { Avatar, Card, Typography } from '@webb-tools/webb-ui-components';
 import assert from 'assert';
 import { FC } from 'react';
-import { SmallChip } from './SmallChip';
+import SmallChip from './SmallChip';
 
 export type ProjectItem = {
   ownerAvatarUrl: string;
