+++ apps/zk-explorer/constants/index.ts
@@ -1,7 +1,6 @@
 export const GITHUB_LANGUAGE_COLORS_API_URL =
   'https://raw.githubusercontent.com/ozh/github-colors/master/colors.json';
 
-export const API_PREFIX = '/api';
 export const ITEMS_PER_PAGE = 12;
 
 export const FEEDBACK_URL =
@@ -14,3 +13,6 @@ export const CONTACT_URL = 'mailto:hello@webb.tools';
 export const IS_DEBUG_MODE = true;
 
 export const PROJECT_DETAIL_CONTAINER_ID = 'project-detail-container';
+
+// TODO: Consider whether HTTP URLs should be allowed.
+export const GITHUB_URL_PREFIX = 'https://github.com/';
