+++ apps/zk-explorer/containers/ProjectHeaderContainer/ProjectHeaderContainer.tsx
@@ -3,7 +3,7 @@ import { Typography } from '@webb-tools/webb-ui-components';
 import cx from 'classnames';
 import GitHubIconWithLink from '../../components/GitHubIconWithLink';
 import { RunCircuitServiceModalTrigger } from '../../components/RunCircuitServiceModalTrigger';
-import { SmallChip } from '../../components/SmallChip';
+import SmallChip from '../../components/SmallChip';
 import SocialChip from '../../components/SocialChip';
 import { MOCK_CIRCUIT_FILE_PATH } from '../../constants/mock';
 import { getProjectHeaderContainerData } from '../../server';
